+++ ci/cloudbuild/builds/cmake-install.sh
@@ -92,6 +92,7 @@ expected_dirs+=(
   ./include/google/cloud/gkehub/v1/configmanagement
   ./include/google/cloud/gkehub/v1/multiclusteringress
   ./include/google/cloud/grpc_utils
+  ./include/google/cloud/gsuiteaddons/v1/
   ./include/google/cloud/internal
   # no RPC services in google/cloud/metastore/logging
   ./include/google/cloud/metastore/logging